# docs/design/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UnitCapabilities": {
        "additionalProperties": false,
        "properties": {
          "case_studies": {
            "items": {
              "$ref": "#/components/schemas/UnitCaseStudy"
            },
            "title": "Case Studies",
            "type": "array"
          },
          "experts": {
            "items": {
              "$ref": "#/components/schemas/UnitExpert"
            },
            "title": "Experts",
            "type": "array"
          },
          "hrm_synced_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hrm Synced At"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "salekit_synced_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salekit Synced At"
          },
          "tech_stack": {
            "examples": [
              [
                "Java",
                "Spring",
                "AWS"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Tech Stack",
            "type": "array"
          }
        },
        "title": "UnitCapabilities",
        "type": "object"
      },
      "UnitCapabilitiesUpdate": {
        "additionalProperties": false,
        "properties": {
          "case_studies": {
            "items": {
              "$ref": "#/components/schemas/UnitCaseStudy"
            },
            "title": "Case Studies",
            "type": "array"
          },
          "experts": {
            "items": {
              "$ref": "#/components/schemas/UnitExpert"
            },
            "title": "Experts",
            "type": "array"
          },
          "hrm_synced_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hrm Synced At"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "salekit_synced_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salekit Synced At"
          },
          "tech_stack": {
            "examples": [
              [
                "Java",
                "Spring",
                "AWS"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Tech Stack",
            "type": "array"
          }
        },
        "title": "UnitCapabilitiesUpdate",
        "type": "object"
      },
      "UnitCaseStudy": {
        "additionalProperties": false,
        "properties": {
          "domain": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "tech_stack": {
            "items": {
              "type": "string"
            },
            "title": "Tech Stack",
            "type": "array"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "title"
        ],
        "title": "UnitCaseStudy",
        "type": "object"
      },
      "UnitContact": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "leader@rikkeisoft.com"
            ],
            "title": "Email"
          },
          "name": {
            "examples": [
              "Nguyen Van A"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "+84 9xx xxx xxx"
            ],
            "title": "Phone"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Delivery Leader"
            ],
            "title": "Title"
          }
        },
        "required": [
          "name"
        ],
        "title": "UnitContact",
        "type": "object"
      },
      "UnitExpert": {
        "additionalProperties": false,
        "properties": {
          "focus_areas": {
            "examples": [
              [
                "Azure",
                "DevOps"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Focus Areas",
            "type": "array"
          },
          "name": {
            "examples": [
              "Sin"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "profile_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          }
        },
        "required": [
          "name"
        ],
        "title": "UnitExpert",
        "type": "object"
      },
      "UnitPublic": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/UnitCapabilities"
          },
          "code": {
            "examples": [
              "D365",
              "AI",
              "JAPAN"
            ],
            "maxLength": 50,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/UnitContact"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "examples": [
              "Rikkei D365 Division"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "inactive"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "created_at",
          "id",
          "code",
          "name",
          "contact",
          "capabilities"
        ],
        "title": "UnitPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Bear In Mind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/units": {
      "get": {
        "description": "Useful for FE dropdowns and admin screens.",
        "operationId": "list_units_api_v1_units_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UnitPublic"
                  },
                  "title": "Response List Units Api V1 Units Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List units",
        "tags": [
          "units"
        ]
      }
    },
    "/api/v1/units/{unit_id}": {
      "get": {
        "operationId": "get_unit_api_v1_units__unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get unit by id",
        "tags": [
          "units"
        ]
      }
    },
    "/api/v1/units/{unit_id}/capabilities": {
      "put": {
        "description": "US3: authenticated unit lead updates capability info; triggers re-embed later.",
        "operationId": "update_capabilities_api_v1_units__unit_id__capabilities_put",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitCapabilitiesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update unit capabilities",
        "tags": [
          "units"
        ]
      }
    },
...
  }
}